seperate the database file
